+++ hipcub/include/hipcub/backend/cub/device/device_select.hpp
@@ -89,6 +89,46 @@ public:
                        stream);
     }
 
+    template<typename IteratorT, typename FlagIterator, typename NumSelectedIteratorT>
+    HIPCUB_RUNTIME_FUNCTION static hipError_t Flagged(void*                d_temp_storage,
+                                                      size_t&              temp_storage_bytes,
+                                                      IteratorT            d_data,
+                                                      FlagIterator         d_flags,
+                                                      NumSelectedIteratorT d_num_selected_out,
+                                                      int                  num_items,
+                                                      hipStream_t          stream = 0)
+    {
+        return hipCUDAErrorTohipError(::cub::DeviceSelect::Flagged(d_temp_storage,
+                                                                   temp_storage_bytes,
+                                                                   d_data,
+                                                                   d_flags,
+
+                                                                   d_num_selected_out,
+                                                                   num_items,
+                                                                   stream));
+    }
+
+    template<typename IteratorT, typename FlagIterator, typename NumSelectedIteratorT>
+    HIPCUB_DETAIL_DEPRECATED_DEBUG_SYNCHRONOUS HIPCUB_RUNTIME_FUNCTION static hipError_t
+        Flagged(void*                d_temp_storage,
+                size_t&              temp_storage_bytes,
+                IteratorT            d_data,
+                FlagIterator         d_flags,
+                NumSelectedIteratorT d_num_selected_out,
+                int                  num_items,
+                hipStream_t          stream,
+                bool                 debug_synchronous)
+    {
+        HIPCUB_DETAIL_RUNTIME_LOG_DEBUG_SYNCHRONOUS();
+        return Flagged(d_temp_storage,
+                       temp_storage_bytes,
+                       d_data,
+                       d_flags,
+                       d_num_selected_out,
+                       num_items,
+                       stream);
+    }
+
     template<typename InputIteratorT,
              typename OutputIteratorT,
              typename NumSelectedIteratorT,
@@ -132,6 +172,45 @@ public:
                   temp_storage_bytes,
                   d_in,
                   d_out,
+                  d_num_selected_out,
+                  num_items,
+                  select_op,
+                  stream);
+    }
+
+    template<typename IteratorT, typename NumSelectedIteratorT, typename SelectOp>
+    HIPCUB_RUNTIME_FUNCTION static hipError_t If(void*                d_temp_storage,
+                                                 size_t&              temp_storage_bytes,
+                                                 IteratorT            d_data,
+                                                 NumSelectedIteratorT d_num_selected_out,
+                                                 int                  num_items,
+                                                 SelectOp             select_op,
+                                                 hipStream_t          stream = 0)
+    {
+        return hipCUDAErrorTohipError(::cub::DeviceSelect::If(d_temp_storage,
+                                                              temp_storage_bytes,
+                                                              d_data,
+                                                              d_num_selected_out,
+                                                              num_items,
+                                                              select_op,
+                                                              stream));
+    }
+
+    template<typename IteratorT, typename NumSelectedIteratorT, typename SelectOp>
+    HIPCUB_DETAIL_DEPRECATED_DEBUG_SYNCHRONOUS HIPCUB_RUNTIME_FUNCTION static hipError_t
+        If(void*                d_temp_storage,
+           size_t&              temp_storage_bytes,
+           IteratorT            d_data,
+           NumSelectedIteratorT d_num_selected_out,
+           int                  num_items,
+           SelectOp             select_op,
+           hipStream_t          stream,
+           bool                 debug_synchronous)
+    {
+        HIPCUB_DETAIL_RUNTIME_LOG_DEBUG_SYNCHRONOUS();
+        return If(d_temp_storage,
+                  temp_storage_bytes,
+                  d_data,
                   d_num_selected_out,
                   num_items,
                   select_op,